requirements.txt
google-cloud-storage
pyspark
Faker
kafka-python
apache-airflow
apache-airflow-providers-ssh
